Clean up Conversation migrator and move it to its own module.
